sla_client/requirements.txt
#pyopengl
PyOpenGL
PyOpenGL_accelerate
numpy
numpy-stl
stl
mercurial
requests
enum
sortedcontainers

